+++ tsconfig.app.json
@@ -20,6 +20,9 @@
       "@/*": [
         "./src/*"
       ],
+      "@filth/*": [
+        "./src/lib/*"
+      ],
       "@helpers/*": [
         "./src/helpers/*"
       ],
