+++ tsconfig.node.json
@@ -23,6 +23,9 @@
       "@/*": [
         "./src/*"
       ],
+      "@filth/*": [
+        "./src/lib/*"
+      ],
       "@helpers/*": [
         "./src/helpers/*"
       ],
